This is a quick visualizer of QPCR data from LightCycler 480 rawfile. It will read the sample names from the raw table, but you also need to provide a .csv to indicate the primer you use for each well.
The output of this script should be two barplot for n = 1 QPCRs, one showing relative exp to internal ctrl, one showing fold change of relative exp to ctrl sample.

For n > 1 QPCRs, replicate number needs to be assigned on sample name, and stripcharts with errorbar (standard deviation) and mean should be made for relative exp to internal ctrl and fold change to control samples.